+++ holberton.h
@@ -55,4 +55,6 @@ char *str_concat(char *, char *);
 
 char *string_nconcat(char *, char *, unsigned int);
 
+int _strlen(char *);
+
 #endif
