--- string_nconcat.c
@@ -1,52 +1,0 @@
-#include <stdio.h>
-#include <stdlib.h>
-/**
- * string_nconcat - concatenates two strings
- * @s1: first string argument
- * @s2: second string argument
- * @n: amount of bytes for s2
- * Return: Null on failure, else ptr to malloc
- */
-
-char *string_nconcat(char *s1, char *s2, unsigned int n)
-{
-	char *ptr;
-	unsigned int i = 0;
-	unsigned int j = 0;
-
-	if (s1 == NULL)
-	{
-		s1 = "";
-	}
-	if (s2 == NULL)
-	{
-		s2 = "";
-	}
-	while (s1[i] != '\0')
-	{
-		i++;
-	}
-	while (s2[j] != '\0')
-	{
-		j++;
-	}
-	if (n > j)
-	{
-		n = j;
-	}
-	ptr = malloc(sizeof(char) * (n + i) + 1);
-	if (ptr == NULL)
-	{
-		return (NULL);
-	}
-	for (i = 0; s1[i] != '\0'; i++)
-	{
-		ptr[i] = s1[i];
-	}
-	for (j = 0; j < n; j++)
-	{
-		ptr[i + j] = s2[j];
-	}
-	ptr[i + j] = '\0';
-	return (ptr);
-}
